browse: Improve revision view with list of changed files and diffs
This commit adds the list of changed files in the revision
view but also the diffs (unified and side-by-side) of each
changed file.

Those new features are reachable from the "Changes" tab
in the revision view. Computation of file diffs will
be requested on the fly as the view is scrolled.

Regarding the diffs, they are only computed for textual
contents and diffs larger than 20 kB will not be computed
by default (the user can still manually request it
if needed).

Closes T921