browse: Change behavior for origin browse url
Url of type /browse/origin/(origin_type)/url/(origin_url) now redirects
to the display of the latest archived snapshot for the given origin
instead of its visits reporting.

Origin visits reporting are now available through the following url
/browse/origin/(origin_type)/url/(origin_url)/visits or by clicking in
the topmost link when browsing the archive in an origin context.

Closes T1049