browse: Support optional parts of persistent identifiers
This commits adds support for optional parts of swh persistent identifiers
in order to add origin context when resolving the swh object to browse.
It is also possible to set lines to highlight for swh content object.

Also to be noted, persistent id urls are now rooted to / instead of /browse.

Closes T1042