persistent identifiers: Resolve optional parts
This commits adds support for optional parts of persistent identifiers
in swh-web in order to add origin context when resolving the swh object
to browse. It is also possible to set lines to highlight for swh content
object.

Also to be noted, persistent id urls are now rooted to / instead of
/browse. Nevertheless backward compatibility is maintained for HAL.

Closes T1042