browse: Improve rendering of reStructuredText README
When a README is written in reStructuredText (rst), convert it to html server
side through the use of pypandoc in order to get a proper display of it.

Indeed, the client side library used to render markdown (showdown.js) does
not handle rst properly and there is no viable solution to convert rst
to html directly in the browser.