browse/vault: Handle dead link for cooked archive
A download link for an archive cooked through the vault (directory or
revision_gitfast) does not remain alive ad vitam aeternam.

As the history of previously cooked objects remains stored in the browser
local storage, we need to handle the case where one wants to download a
previously cooked archive from a dead link.

In that case, check the link availability first before proceeding to the
real download. If it is dead, display a modal asking the user if he wants
to recook the archive.

Closes T1082