api: Implement endpoint for save origin feature
This commit adds a new endpoint to the Software Heritage Web API in order
to offer users the possibility to submit software origins to save into
the archive.

The type of origins that can be currently saved are: git, hg and svn.
Submitted urls are first test again a blacklist/whitelist to determine
if they are authorized to be loaded or not. Unknown urls will be put
under pending state until a manual review is performed to determine if
they can be safely loaded or not.

Related T336
Related T1121