tests: Add module to generate relevant data as tests input
In order to avoid harcoding tests input data and get closer to real world ones,
populate a test archive by loading in it a couple of lightweight git repositories.

The ids of the objects in this test archive (contents, directories, revisions, ...)
will then be provided as tests input in order to retrieve their associated data
from the in-memory storages. Proceeding like this will allow us to remove a
lot of mocks in the tests implementation.

Related T1271