common.origin_save: Force status update of tasks still marked as scheduled
Some submitted origin save tasks were still marked as scheduled in production.
This may happen when the celery tasks listener is down, preventing the update
of the tasks status in scheduler database.

So check if the tasks have been executed or not, and update their status in
the swh-web database.