assets/templates: Switch back to classical links in table rows
Revert modifications introduced in 1dab1b07c097 as it was not a good idea
thinking it back. In particular, navigating the webapp using the keyboard
was not optimal at all.

Also reinstate the link outlines in order to clearly see which link is
currently focused when navigating through the keyboard.