debian/control: Explicitely add python3-magic in Depends section
python3-magic corresponds to file-magic in stretch while it corresponds to python-magic
starting buster.

As swh-web now favorizes the use of python-magic (0721db0ce7e8), we need to explicitely
add python3-magic in runtime dependencies as dh_python3 will not be able to automatically
add this dependency for debian stretch.