assets/webapp-utils: Implement 'Jump to page' UI for DataTables
When clicking on an ellipsis in the page number links below a DataTable,
a popover will be displayed with a select element in it to jump to
a specific page number.

Apply this UI improvement to the 'Save code now' requests table as it
now contains more than 4000 rows in it.