admin/origin-save: Display save request task info on admin user demand
Add an info button in a new column of the 'Save code now' accepted requests list.

By clicking on it, a popover will be displayed while a fetch request will be sent
to the backend to get the loading task execution info (if available) associated
to the request.

Once the response of the fetch request is received, task execution info data will
be formatted and displayed in a table.

Closes T1976