browse: Fix release browsing issue in snapshot context views
Sometimes, when one tries to browse objects releated to a specific release contained
in a snapshot, a 404 error is raised while the release exists.

The issue was due to the associated snapshot content being truncated to 1000 branches
when the webapp retrieves it.

Error handling when ones tries to browse an invalid release has also been improved.

Closes T2108