Move static folder out of the Python tree
It does not contain Python files but rather Javascript and CSS source files
generated by webpack but also fonts and image files. So it seems a better
practice to move this folder at the root of the repository and package it
as module data files when creating a distribution wheel.

Closes T2155