apidoc: Generate a dedicated doc view for each decorated endpoint
In order to simplify the use of the apidoc module, create a dedicated view
to display HTML documentation for each decorated endpoint.

This fixes an issue when an endpoint only accepts POST requests.

The noargs parameter is still used but is no more mandatory to an endpoint
without URL arguments. It simply enables to avoid redirecting to the endpoint
doc view from the endpoints list view as response data can be immediately
displayed.