api: Update /revision/(sha1_git)/log/ endpoint
Clarify the order in which revisions are returned (BFS) in documentation.

Stop returning next-link in response headers as it is not sufficient to properly
resume the BFS traversal on the revision graph.

Remove any references to pagination in documentation and rename per_page query
parameter to limit (also set its max value to 1000).

Add warning in documentation explaining how a client can obtain the full BFS
traversal of the revision graph.

Remove useless optional prev_sha1s URL argument.

Closes T2450