templates/layout: Add a switch to toggle full width rendering
Add a switch in the top left part of the UI in order to toggle full width
rendering of web application content.

This can be useful for high resolution screens or to extend the viewport width
for source code rendering.

The full width rendering state is also saved in browser local storage in order
to restore it on each web application page load.

Closes T2456