common/identifiers: Fix missing lines qualifier processing in resolver
When a content object is anchored to a revision, the resolved url from the
associated SWHID will target the revision view instead of the content one.

So ensure the SWHID lines qualifier is also processed in that case in order
for the "scroll to lines" feature to work each time a content is displayed
in a view.

Closes T2464