templates/origin-save: Add a new "Info" column in save requests table
An "Info" button has been added in that new column. When clicking on it, a
popover containing various loading task information will be displayed.

For anonymous users, the displayed info are:

  - Task type
  - Task arguments
  - Task scheduling date
  - Task start date
  - Task termination date
  - Task duration
  - Task log (with Traceback removed in case of failed task,
    only the exception is displayed)

For staff users, more internal info and a full traceback are displayed in
case of a failed task.

Related to T2325