templates/origin-save: Add a "Save again" button in requests list
When clicking on it, the save request submission form will be automatically
filled with the origin info related to the associated requests table row
and the view will be scrolled to the top.

Also rename "Visit type" to "Origin type" ast it is more comprehensible
for non swh staff users.

Closes T2325