origin_visits/get_origin_visit: Improve default visit picking strategy
When no visit info (timestamp, id or snapshot) is provided to that function
return the last full visit with a valid snapshot or the last partial visit
with a valid snapshot otherwise.

This fixes a browsing issue when an origin is no more available but swh keeps
visiting it afterwards.

Use new swh.storage.algos.origin.origin_get_latest_visit_status utility
function in the revised get_origin_visit_implementation.