assets/diff-utils: Add diff lines highlighting feature
That new feature enables to highlight code fragments in a diff generated
from the revision view by selecting a lines range with the mouse (click
and Shift+click on line numbers).

Both unified and split diff can be highlighted. For the split case,
only lines from the left or right part can be highlighted but also lines
spanning both.

When a diff gets highlighted, an URL fragment will be added describing the
highlighting state. When the revision view gets reloaded with it, the
diff will be automatically computed, highlighted and displayed.

Closes T2491