SWHIDs: Fix URLs escaping when a qualifier contains escaped chars
When a SWHID qualifier contains special characters like '=' or ';', it
must be escaped in order for the SWHID to be properly parsed afterwards.

When using such a SWHID as part of an URL argument, it must also be escaped
for special characters in order for a browser to resolve the proper URL.

That commit fixes invalid unescaped URLs that were previously inserted in
some webapp pages (leading to 400 errors).

Closes T2641