api: Add /graph endpoint proxying Software Heritage Graph service
Add a new /graph endpoint to the Web API acting as a proxy for the Software
Heritage Graph service. It forwards graph queries to it and return its raw
responses.

The endpoint requires authentication and user special permission in order
to be able to query it.

It also adds an extra query parameter "resolve_origins" enabling to resolve
origin urls from their sha1 representations returned by the Graph service.

Closes T2589