auth: Add user profile UI for managing account and bearer tokens
Add a new view enabling for a user to manage its account (email, password)
but also its bearer tokens for Web API authentication.

Account details are displayed in the view but edition of personal information
must be done through the dedicated Keycloak account management UI (link to it
is available in the view).

The view is reachable by clicking on the username on the top right part of
the Web UI once authenticated.

Closes T2718