vault: Fix vault response schema
Names got changed in the vault api backend. Unfortunately, this passed
undetected both by the vault and the webapp tests.

Furthermore, as the try catch within the js layer is silent, this fails
silently.

This commit fixes the situation by mapping from the new names to the old names.
This keeps the the api behavior consistent.