package.json: Remove duplicated postcss configuration
The postcss configuration is already defined in webpack configuration so
it can be removed from the package.json file.

It was needed at some time due to a postcss configuration bug but this is
not the case anymore.

It also enables to remove disturbing stylelint warnings when compiling
with webpack.