common/archive: Handle single slash after protocol in lookup_origin
Some external resolver or software might mangle the "://" character sequence
into ":/" for the origin qualifier value of a SWHID.

This can lead to 404 errors when trying to resolve such a SWHID as the origin
will be found missing due to the typo.

So handle that special case to avoid such errors.

Related to T2797

Closes T2878