common: Log traceback of exception caught in view error handler
Traceback of an exception caught in a django view error handler was included
in error responses but not logged to stderr.

While the exception traceback is clearly readable in browse error pages, it
is not in Web API responses because of the JSON format.

So ensure the traceback is also logged to stderr for better debugging.