cypress: Migrate from cy.route API to cy.intercept API
cy.route has been deprecated in cypress 6.0 in favor of cy.intercept
and will be removed soon.

So migrate code stubbing network requests to the intercept API.

The only drawback of the intercept API is that currently stubbed responses
can not be overridden but hopefully a simple workaround can be used
(see https://github.com/cypress-io/cypress/issues/9302).