webpack/weblabels-plugin: Fix Web Labels table in production mode
Recent changes in webpack modified the content of the JSON stats object
when compiling assets in production mode. Notably a module chunk is now
identified by an integer instead of its name.

So add an extra mapping between a chunk id and its name in order to fix
the generated Web Labels table.