browse: Filter out pull request branches in snapshot content
Github origins might contain a lot of branches corresponding to pull
requests in their associated snapshots.

Those branches are not really of interest and add a lot of noise in
the branches views of swh-web.

So filter out these branches from snapshots by default in the browse
web application in order to improve the display.

Nevertheless, those branches are not filtered out when querying a
snapshot with the Web API.

Related to T2782