auth: Use generic Django authentication backends from swh-auth
Django authentication backends and middleware based on the use of
Keycloak and OpenID Connect have been moved to the swh-auth module
and  turned into generic ones in order to easily reuse them in any
Django application.

So remove all code and tests that have been moved in swh-auth and
configure swh-web to use the generic authentication backends and
middleware.

Closes T3150