common/origin_save: Prevent erroneous save code now status overriding
Loading status for a save code now request is first retrieved from the
scheduler database.

When a visit date for a save code now request has been found, the visit
status is then checked to possibly update the final status of the
request to failed (when an origin is not found for instance).

However to avoid sending too many costly queries to the storage,
the visit date and status are only checked once and stored to
swh-web database.

So ensure to not override a final loading task status (succeeded or
failed) to avoid invalid save code now request reporting.