Separate save code now status refresh routine from the listing ui
This introduces a new endpoint to refresh the status of the save code now requests. This
endpoint will be called periodically. That concern is out of scope of the listing ui and
source of status update desynchronization [1]

The listing ui now feels more immediate as implementation wise, it no longer writes when
actually listing rows.

This slightly improves the code by typing it a bit more. A new SaveOriginRequestInfo got
added to represent the save origin request to display view side. This matches what
other views manipulates.

[1] T3278#63827

Related to T3280