origin_save: Use API endpoint to submit save request in Web UI form
The "Save code now" Web UI form was using an internal dedicated endpoint
to submit an origin save request.

That endpoint is redundant with the /origin/save Web API one so drop
it and use Web API instead.

The internal endpoint was there to validate the CSRF token injected
by Django but the /origin/save already has a rate limit of 10 POST
requests per hour to prevent abuse.