Save code now: Check origin exist prior to save request
This slightly modifies the current save code now workflow. Prior to register a new
origin url for "save code now", this first ensures said origin exists (through a HEAD
request done server side).

If the origin does not exist, an explanation notification is provided to the user.
Otherwise, the origin is submitted for registration as usual.

This should decrease the number of loading visits ending up in not_found status. This
also prepares work for the incoming new bundle visit type (which needs both that check
and the HEAD request anyway to fetch some more information).

Related to T3213