assets/webapp-utils: Add lowercase validator for core SWHIDs
When entering a SWHID in a search input, check its core part is in lowercase
and do not validate the form if it is not.

An error message clearly indicating that core SWHID should be in lowercase
will then be displayed.

Nevertheless, core SWHID with uppercase chars in it will still be automatically
fixed when provided to its resolve URL /<swhid>

Related to T3298