common/origin_save: Update missing information when available
Due to some internal details [1], it happens that save code now requests are updated
with missing information. Broadening a bit the filtering on the save code now requests
to list for update, we can try and update those information more frequently (even when
the main information is already updated). This way, the save code now requests will
eventually be updated fully when the information lateness subsides.

[1] Replication lag

Related to T3346