assets: Upgrade highlight.js to version 11.x
This new major version of highlight.js brings notable improvements in
terms of language parsing.

As the github theme we were using has been updated to match the current
github.com one, we now use a custom theme adapted from the old github
one to keep the same look and feel for code highlighting as before.

Closes T3368