Simplify save code now request status updates using visit statuses
This maximizes the use of the origin visit status to determine the save code now request
final status (succeeded or failed). When such data is not yet available, the task is
probably not yet scheduled, scheduled or currently running. So, it fallbacks to use the
scheduler task/task_run to deduce such non-final statuses (e.g scheduled/running/not yet
scheduled/...).

Computing the save code now status that way will decrease the potential discrepancy [1]
between the status displayed and the origin not being browsable yet.

[1] Most likely due to some replication lag

Related to T3378