Search inside the origin releases with name
This change enables the possibility to search with release names inside
the browse origin releases page.A search textbox is added, which is
connected to a JS function and that reloads the page with the
release_name_include query param.
Filtering with the name is happening at the storage level.

Related to T3404