browse/snapshot_context: Handle empty revisions history in log view
It now exists in the archive snapshot containing only releases targeting
directories (typically those generated by the package loaders).

Thus these snapshots do not contain any revisions, this was leading to a
server error when trying to browse their history view.

Now a 404 error will be returned when trying to browse the revisions
history of such snapshots.

Related to T3840