browse/snapshot_context: Fix revision_found setting in directory view
That variable was set in an incorrect way. As revision_id can be None,
it should be initialized to False and set to True after checking the
revision can be found in the archive.

This fixes the non display of the "History" button in the snapshot
directory view when a snapshot contains no revisions.

Related to T3840