auth: Add UserMailmap model and edition endpoints
Add a new database table to store mailmap settings that will be
used to add display names for an authors that can be found in the
archive.

Add endpoints to add or update these mailmap settings restricted
to staff users or users with special permission.
They will be used in the upcoming Web UI to manage mailmaps.