admin: Add mailmaps administration Web UI
Add a new administration view to manage the mailmaps in the
archive.

The view allows to add new mailmaps or update existing ones.

It requires a special user permission "swh.web.admin.mailmap"
to be used.

Only mailmaps created by users with that permission will be
displayed in that view.