Add a sync_mailmaps management command
The command refreshes all the active mailmaps, and disables every
mailmap that has recently been updated and marked as disabled, once.

To do these actions as atomically as possible, we make sure to select
all rows for update, and to defer all _date fields in the default
queryset manager for UserMailmap objects to avoid resetting them to
earlier values by mistake.