templates/origin-visits: Prefer to show all visits by default
There is a lot of origins without any full visit in the archive so in
that case the origin visits page looks empty when reaching it as the
default is to display full visits with different snapshots.

So prefer to display all visits by default and disable the full visits
switch radio buttons when there is no such visits.