requirements: Workaround another ImportError when running pytest
When calling "./bin/install" from swh-environment in a freshly created
venv, for some reason pip will end up installing flask==1.1.2 when
processing swh-web (likely related to the venv state after installing
other swh modules).

That flask version is quite outdated and will fail to import the json
module from the version of itsdangerous installed in the venv.

So as a workaround pin flask to the first version that does not trigger
that ImportError issue.