deposit moderation view: Improve deposit types visualization
This explicits the deposit's type, either 'code' or 'meta'. The uri column is now filled
depending on that type. Mostly, the parsing of the deposit's raw xml content is done to
try and extract that potentially provided information.

If not found, this uses the same heuristics as before (either use the origin_url if
provided or extract the origin out of the swhid_context).

Related to T3680