api/add_forge_now: Allow to use requests list endpoint with datatables
datatables is the javascript library we use on the frontend side to
display interactive tables.

As we use server-side processing, table data must be provided in a
paginated way by an HTTP endpoint.

Response format expected by datatables is different from the one
returned by the Web API endpoint listing add-forge requests.

So adapt the response format of that endpoint when we know the
input request has been sent by datatables.

Related to T3989
Related to T3991