add_forge_now: Move datatables listing in dedicated endpoint
Merging datatables special processing in /api/1/add-forge/request/list
endpoint implementation was quite disturbing.

So better using a dedicated endpoint for that in a similar manner as
with save code now requests listing.

The endpoint to use with datatables is now the following one:
/add-forge/request/list/datatables.

Related to T3989
Related to T3991