cypress: Add missing global intercept for https://status.swh.org
Each time a swh-web HTML page is loaded, an HTTP request is sent to
https://status.softwareheritage.org for getting the data needed to
display the status widget in the top right of the Web UI.

That request was not globally intercepted by cypress resulting in
making real HTTP requests on the internet while executing tests.
As a consequence, this was slowing down the whole testing process.

So ensure that request is globally intercepted to speedup tests
execution.