cypress: Mock costly HTTP request to /coverage endpoint
Before each test suite, cypress visits the root page to get some
references on javascript objects used in tests.

However the root page requests the costly /coverage endpoint to
render the archive coverage widget.

We do not need that widget to render for cypress tests so better
mocking that request to an empty one to speedup each test suite
startup.