assets/save: Fix Web UI tabs switch when navigating in browser history
In the Save Code Now Web UI, the two displayed tabs have different URLs
but it was not possible to navigate back and forth between them when
clicking on the "Go back" and "Go forward" browser buttons.

So ensure the correct tab will be active when URL fragment has changed.