django: Ensure defined URLs have trailing slash
URLs targeting a django view must have a trailing slash as
django builtin feature APPEND_SLASH is activated by default.

If an URL definition does not have a trailing slash, this
will  result in a 404 error as django will automatically add
a trailing slash if a user forgot to add it and thus the URL
resolving will fail.