Filter user add-forge-now requests when authenticated
This introduces a checkbox to allow users to filter their own requests. It's turned on
by default. The checkbox is not shown when connected anonymously.

Either browsing anonymously or unchecking the box (when connected) allows to browse all
requests.

This also unifies the rendering with the save code now view:
- same datatable layout
- the spinner logo during computation

Related to T4011