admin/deposit: Use new deposit private API endpoint to list deposits
Previously, the admin-deposit-list view was retrieving all deposits data
using the /1/private/deposits/ swh-deposit endpoint and was filtering
those client side.

That approach was not efficient at all as the number of deposits
increased.

So from now on, swh-web will simply forward the listing requests to
the new /1/private/deposits/datatables/ endpoint of swh-deposit which
will perform the same filtering and pagination operations as before in
an efficient way.

Also remove code that is now dead and ensure deposit private API base
URL can be set in configuration with or without trailing slash.

Related to T3128