cypress: Use non github origin for save code now tests
Cypress tests for save code now was using https://github.com/memononen/libtess2
as input origin for most of these tests. But using a github origin will trigger
an HTTP request to https://api.github.com/repos/ client-side to get the canonical
URL of the repository.

That request was not intercepted by cypress so real HTTP requests on the internet
were issued during the tests which slowdown them and can make them fail under
certain conditions.

So prefer to use a sample git origin URL not related to github to avoid these kind
of issues.