api/raw: Restrict raw object endpoint use to privileged users
As that endpoint can leak sensible information regarding authors in
the archive, it has been decided to not make it publicly available.

So restrict its use to staff users and users with new permission
"swh.web.api.raw_object".

Also lift rate limit on that new endpoint to ease development
of the bridge of Software Heritage to IPFS.

Related to T4226