misc/coverage: Add a focus query parameter to emphasize origin types
When that parameter is used, the lists of origin types it contains will
be emphasized in the archive coverage view by adding a red border around
the associated bootstrap cards.

Cards content will also be expanded by default and no maximum height
will be applied.

The purpose is to emphasize the result of a given lister. Such links
will be made available in the documentation page compiling the current
statuses of available listers.

Related to T3117