misc/coverage: Improve display of origin counters in production/staging
If the global counter for an origin type is equal to zero, do not display
the bootstrap card for that origin type.

If a visit type counter for an origin type is equal to zero, do not
display it in the counters table for that origin type.

It enables to not take into account origins that have been listed
but not loaded without having to hardcode that information in
coverage widget implementation.

Also use origins_enabled field from scheduler metrics instead of
origins_known one to compute the number of loaded origins per
visit type.

Related to T3117